feat: move `withNextApp` and `withNextDocument` to `@helpwave/common`
this requires using the same Twind instance everywhere, therefore all
imports had to be updated and `@helpwave/common` has to export its own
instance to share it with client applications.
