+++ app/pages/_app.tsx
@@ -1,9 +1,9 @@
 import Head from 'next/head'
 import type { AppProps } from 'next/app'
 import { Inter, Space_Grotesk as SpaceGrotesk } from '@next/font/google'
-import { tw } from '@twind/core'
+import { tw } from '@helpwave/common/twind/index'
 import { ProvideLanguage } from '@helpwave/common/hooks/useLanguage'
-import withNextApp from '../twind/next/app'
+import withNextApp from '@helpwave/common/twind/next/app'
 import { config } from '@helpwave/common/twind/config'
 
 const inter = Inter({
