+++ app/pages/profile.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 import type { PropsWithChildren } from 'react'
 import type { NextPage } from 'next'
 import Head from 'next/head'
-import { tw } from '@twind/core'
+import { tw } from '@helpwave/common/twind/index'
 import { useAuth } from '../hooks/useAuth'
 import { useRouter } from 'next/router'
 import { Header } from '../components/Header'
