+++ landing-page/pages/_app.tsx
@@ -2,8 +2,8 @@ import Head from 'next/head'
 import type { AppProps } from 'next/app'
 import { Toaster } from 'react-hot-toast'
 import { Inter, Space_Grotesk as SpaceGrotesk } from '@next/font/google'
-import { tw } from '@twind/core'
-import withNextApp from '../twind/next/app'
+import { tw } from '@helpwave/common/twind/index'
+import withNextApp from '@helpwave/common/twind/next/app'
 import { config } from '@helpwave/common/twind/config'
 import { ProvideLanguage } from '@helpwave/common/hooks/useLanguage'
